+++ src/main/java/com/network/Pilotosformulauno/Servicio/IServicio_Escuderia.java
@@ -2,6 +2,6 @@ package com.network.Pilotosformulauno.Servicio;
 
 import com.network.Pilotosformulauno.Dominio.Escuderia;
 
-public interface IServicio_Escuderia extends ICrud<Escuderia,Integer> {
+public interface IServicio_Escuderia extends ICrud<Escuderia,Long> {
 
 }
