+++ src/main/java/com/network/Pilotosformulauno/Servicio/IServicio_Piloto.java
@@ -2,7 +2,7 @@ package com.network.Pilotosformulauno.Servicio;
 
 import com.network.Pilotosformulauno.Dominio.Piloto;
 
-public interface IServicio_Piloto extends ICrud<Piloto,Integer> {
+public interface IServicio_Piloto extends ICrud<Piloto,Long> {
 
 
 }
